refactor: Reduce code redundancy (for mapping syscall nrs to names and vice versa); use long for syscall nrs

+++ src/cli.h
@@ -11,7 +11,7 @@ typedef struct cli_args {
     bool list_syscalls;
     pid_t attach_to_process;
     bool follow_fork;
-    int pause_on_scall_nr;
+    long pause_on_scall_nr;
 #ifdef WITH_STACK_UNWINDING
     bool print_stack_traces;
 #endif /* WITH_STACK_UNWINDING */
