+++ src/trace/tracing.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 
-int do_tracer(pid_t pid, bool attach_to_tracee, int pause_on_syscall_nr, bool follow_fork
+int do_tracer(pid_t pid, bool attach_to_tracee, long pause_on_syscall_nr, bool follow_fork
 #ifdef WITH_STACK_UNWINDING
             , bool print_stacktrace
 #endif /* WITH_STACK_UNWINDING */
